Fix empty terrain scale

+++ Assets/Scripts/Ozone SCMAP Code/ScmapEditor.cs
@@ -168,7 +168,7 @@ public partial class ScmapEditor : MonoBehaviour
 
 
 		WaterLevel.transform.localScale = new Vector3(HalfxRes, 1, HalfzRes);
-		TerrainMaterial.SetFloat("TerrainScale", 1 / xRes);
+		TerrainMaterial.SetFloat("TerrainScale", (float) 1.0 / xRes);
 		TerrainMaterial.SetFloat("_GridScale", HalfxRes);
 		TerrainMaterial.SetTexture("UtilitySamplerC", map.UncompressedWatermapTex);
 		WaterMaterial.SetFloat("_GridScale", HalfxRes);
